+++ R/LMMsolve-class.R
@@ -118,12 +118,12 @@ LMMsolveObject <- function(logL,
 #' on which.
 #'
 #' @examples
-#' ## Fit model on john.alpha data from agridat package.
-#' data(john.alpha, package = "agridat")
+#' ## Fit model on oats data.
+#' data(oats.data)
 #'
 #' ## Fit simple model with only fixed effects.
 #' LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
-#'                 data = john.alpha)
+#'                 data = oats.data)
 #'
 #' ## Obtain table of effective dimensions.
 #' summ1 <- summary(LMM1)
@@ -189,12 +189,12 @@ print.summary.LMMsolve <- function(x,
 #' mixed model.
 #'
 #' @examples
-#' ## Fit model on john.alpha data from agridat package.
-#' data(john.alpha, package = "agridat")
+#' ## Fit model on oats data
+#' data(oats.data)
 #'
 #' ## Fit simple model with only fixed effects.
 #' LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
-#'                 data = john.alpha)
+#'                 data = oats.data)
 #'
 #' ## Obtain coefficients.
 #' coefs1 <- coef(LMM1)
@@ -250,12 +250,12 @@ coef.LMMsolve <- function(object,
 #' @returns A vector of fitted values.
 #'
 #' @examples
-#' ## Fit model on john.alpha data from agridat package.
-#' data(john.alpha, package = "agridat")
+#' ## Fit model on oats data
+#' data(oats.data)
 #'
 #' ## Fit simple model with only fixed effects.
 #' LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
-#'                 data = john.alpha)
+#'                 data = oats.data)
 #'
 #' ## Obtain fitted values.
 #' fitted1 <- fitted(LMM1)
@@ -275,12 +275,12 @@ fitted.LMMsolve <- function(object,
 #' @returns A vector of residuals.
 #'
 #' @examples
-#' ## Fit model on john.alpha data from agridat package.
-#' data(john.alpha, package = "agridat")
+#' ## Fit model on oats.data
+#' data(oats.data)
 #'
 #' ## Fit simple model with only fixed effects.
 #' LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
-#'                 data = john.alpha)
+#'                 data = oats.data)
 #'
 #' ## Obtain fitted values.
 #' residuals1 <- residuals(LMM1)
@@ -304,12 +304,12 @@ residuals.LMMsolve <- function(object,
 #' @returns The restricted maximum log-likelihood of the fitted model.
 #'
 #' @examples
-#' ## Fit model on john.alpha data from agridat package.
-#' data(john.alpha, package = "agridat")
+#' ## Fit model on oats data
+#' data(oats.data)
 #'
 #' ## Fit simple model with only fixed effects.
 #' LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
-#'                 data = john.alpha)
+#'                 data = oats.data)
 #'
 #' ## Obtain log-likelihood.
 #' logLik(LMM1)
@@ -341,12 +341,12 @@ logLik.LMMsolve <- function(object,
 #' @returns The deviance of the fitted model.
 #'
 #' @examples
-#' ## Fit model on john.alpha data from agridat package.
-#' data(john.alpha, package = "agridat")
+#' ## Fit model on oats.data
+#' data(oats.data)
 #'
 #' ## Fit simple model with only fixed effects.
 #' LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
-#'                 data = john.alpha)
+#'                 data = oats.data)
 #'
 #' ## Obtain deviance.
 #' deviance(LMM1)
@@ -449,6 +449,7 @@ predict.LMMsolve <- function(object,
                    function(x) {attr(x,which="termType") == "grp"}))
   if (s2 > 0) stop("predict function for grp() not implemented yet")
 
+  # check argument derivative
   if (!is.null(deriv)) {
     if (!is.character(deriv)) {
       stop("Argument deriv should be a character string\n")
@@ -459,45 +460,15 @@ predict.LMMsolve <- function(object,
     if (object$family$family != "gaussian") {
       stop("Derivatives for non-gaussian data not implemented yet\n")
     }
-
-    isDeriv <- sapply(object$splRes,FUN= function(spl) {
-                                        nameVar <- names(spl$x)
-                                        if (length(nameVar) > 1) return(FALSE)
-                                        nameVar == deriv})
-    if (all(!isDeriv)) stop("Cannot find derivative for ", deriv, "\n")
-    spl_nr <- which(isDeriv)
-
-    spl <- object$splRes[[spl_nr]]
-    chkValBsplines(spl, newdata)
-    knots <- spl$knots[[spl_nr]]
-    pord <- spl$pord
-    scaleX <- spl$scaleX
-    x <- newdata[[deriv]]
-    dB <- Bsplines(knots, x, deriv = 1)
-    nRow <- nrow(newdata)
-    dim <- object$dim
-    lU <- list()
-    for (i in seq_along(dim)) {
-      lU[[i]] <- spam::spam(x = 0, nrow = nRow, ncol = dim[i])
+    # at the moment assuming 1 spline component:
+    if (length(object$splRes) > 1) {
+      stop("Derivatives for multiple splines not implemented yet\n")
+    } else {
+      # for that single component, do a check:
+      if (!(deriv %in% names(object$splRes[[1]]$x))) {
+        stop("Cannot find derivative for ", deriv, "\n")
+      }
     }
-
-    labels <- c(object$term.labels.f, object$term.labels.r)
-    ndx.r <- which(spl$term.labels.r == labels)
-    if (!is.null(spl$term.labels.f)) {
-      ndx.f <- which(spl$term.labels.f == labels)
-      G <- constructG(knots = knots, scaleX = scaleX, pord = pord)
-      dBG <- dB %*% G
-      dBG <- dBG[,-1, drop=FALSE]
-      lU[[ndx.f]] <- dBG
-    }
-    lU[[ndx.r]] <- dB
-    U <- Reduce(spam::cbind.spam, lU)
-    outDat <- newdata
-    outDat[["ypred"]] <- as.vector(U %*% object$coefMME)
-    if (se.fit) {
-      outDat[["se"]] <- calcStandardErrors(C = object$C, D = U)
-    }
-    return(outDat)
   }
 
   splFixLab <- sapply(object$splRes, function(x) { x$term.labels.f })
@@ -512,9 +483,15 @@ predict.LMMsolve <- function(object,
     ## check whether the values are in range Bsplines
     chkValBsplines(spl, newdata)
     x <- spl$x
+    ## set the IsDeriv boolean 0/1 for Bsplines argument
+    if (is.null(deriv)) {
+      isDeriv <- rep(0, length(x))
+    } else {
+      isDeriv <- 1*(names(x) == deriv)
+    }
     xGrid[[s]] <- lapply(X = seq_along(x), FUN = function(i) {
       newdata[[names(x)[i]]]})
-    Bx <- mapply(FUN = Bsplines, spl$knots, xGrid[[s]], deriv=0)
+    Bx <- mapply(FUN = Bsplines, spl$knots, xGrid[[s]], deriv=isDeriv)
     BxTot[[s]] <- Reduce(RowKronecker, Bx)
     G <- lapply(X=spl$knots, FUN = function(x) {
       constructG(knots = x, scaleX = spl$scaleX, pord = spl$pord)})
@@ -539,7 +516,11 @@ predict.LMMsolve <- function(object,
     lU[[i]] <- spam::spam(x = 0, nrow = nRow, ncol = dim[i]/nCat)
   }
   # intercept:
-  lU[[1]] <- spam::spam(x = 1, nrow = nRow, ncol = 1)
+  if (is.null(deriv)) {
+    lU[[1]] <- spam::spam(x = 1, nrow = nRow, ncol = 1)
+  } else {
+    lU[[1]] <- spam::spam(x = 0, nrow = nRow, ncol = 1)
+  }
 
   labels <- c(object$term.labels.f, object$term.labels.r)