+++ R/LMMsolve.R
@@ -59,22 +59,22 @@
 #' this object.
 #'
 #' @examples
-#' ## Fit models on john.alpha data from agridat package.
-#' data(john.alpha, package = "agridat")
+#' ## Fit models on oats.data
+#' data(oats.data)
 #'
 #' ## Fit simple model with only fixed effects.
 #' LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
-#'                 data = john.alpha)
+#'                 data = oats.data)
 #'
 #' ## Fit the same model with genotype as random effect.
 #' LMM1_rand <- LMMsolve(fixed = yield ~ rep,
 #'                      random = ~gen,
-#'                      data = john.alpha)
+#'                      data = oats.data)
 #'
 #' ## Fit the model with a 1-dimensional spline at the plot level.
 #' LMM1_spline <- LMMsolve(fixed = yield ~ rep + gen,
 #'                        spline = ~spl1D(x = plot, nseg = 20),
-#'                        data = john.alpha)
+#'                        data = oats.data)
 #'
 #' ## Fit models on multipop data included in the package.
 #' data(multipop)
@@ -263,6 +263,10 @@ LMMsolve <- function(fixed,
       splResList[[i]] <- splRes
       ## Add to design matrix fixed effect X.
       X <- cbind(X, splRes$X)
+      ## Check whether adding the fixed part gives a singularity.
+      if (qr(X)$rank < ncol(X)) {
+        stop("singularity problem ", splRes$term.labels.f, " in spline argument\n")
+      }
       ## Add to design matrix random effect Z.
       Z <- spam::cbind.spam(Z, splRes$Z)
       ## Expand matrices Ginv to the updated Z.