+++ inst/tinytest/test_LMMsolve.R
@@ -5,6 +5,12 @@ load("testdata.rda")
 expect_error(LMMsolve(fixed = pheno ~ cross, data = "testDat"),
              "data should be a data.frame")
 
+expect_error(LMMsolve(fixed= pheno~ cross, data=testDat, theta=c(1,1,1)),
+             "Argument theta has wrong length")
+
+expect_error(LMMsolve(fixed= pheno~ cross, data=testDat, grpTheta=c(1,1,1)),
+             "Argument grpTheta has wrong length")
+
 ## Test fixed, random and residual formulas (spline part is tested in spl checks).
 
 expect_error(LMMsolve(fixed = ~ cross, data = testDat),
@@ -89,8 +95,22 @@ expect_error(LMMsolve(fixed = pheno ~ cross, random = ~grp(QTL),
                       group = list(QTL = 3:5), theta=c(-1,1), data = testDat),
              "theta should have positive values.")
 
+# Test for grpTheta
+expect_error(LMMsolve(fixed = pheno ~ cross, random = ~grp(QTL),
+         group = list(QTL = 3:5), grpTheta=c("a","b"), data = testDat),
+         "grpTheta should be integers 1,2,...nGrp")
 
+# Test for offset to be numeric
+expect_error(LMMsolve(fixed = pheno ~ cross, random = ~grp(QTL),
+                      group = list(QTL = 3:5), offset = "cross", data = testDat),
+             "offset should be numeric")
 
+# Test for family argument
+tst_fam <- function() {return(0) }
+expect_error(LMMsolve(fixed = pheno ~ cross, random = ~grp(QTL),
+                      group = list(QTL = 3:5), data = testDat,
+                      family = tst_fam()),
+             "argument family not correct")
 
 ## Test for syntax names:
 dat <- data.frame(x = 1:3, `a-b` = 3:1, `a+b` = 1:3, ab = 3:1, check.names = FALSE)