+++ inst/tinytest/test_predict.R
@@ -44,6 +44,10 @@ newdat2 <- data.frame(x= seq(-1,1, length = 10))
 expect_error(predict(obj, newdata = newdat2, se.fit = TRUE, deriv="x"),
             "Variable x outside range of B-splines basis")
 
+newdat3 <- data.frame(z = seq(-1,1, length = 10))
+expect_error(predict(obj, newdata=newdat3),
+             "variables (x) in data.frame newdata missing.", fixed=TRUE)
+
 # multinomial example
 set.seed(1234)
 
@@ -84,3 +88,53 @@ expect_error(predict(obj, newdata = newdat, se.fit=TRUE),
 
 pred5 <- predict(obj, newdata = newdat, se.fit=FALSE)
 expect_equivalent_to_reference(pred5, "pred6")
+
+# make predictions for oats.data:
+data(oats.data)
+obj7 <- LMMsolve(yield ~ rep + gen,
+                 random = ~block:rep,
+                 data = oats.data)
+newdat7 <- data.frame(rep = "R1", gen = levels(oats.data$gen))
+pred7 <- predict(obj7, newdata = newdat7)
+expect_equivalent_to_reference(pred7, "pred7")
+
+# test if all fixed effects are defined in newdata:
+newdat8 <- data.frame(gen = "G11")
+expect_error(predict(obj7, newdata=newdat8),
+                     "variables (rep) in data.frame newdata missing.", fixed=TRUE)
+
+# example deriv for two dimensional splines
+
+set.seed(1234)
+f <- function(x,y) { x^2*y + 3*y^2*x}
+df_dx <- function(x,y) { 2*x*y + 3*y^2}
+df_dy <- function(x,y) { x^2 + 6*x*y }
+
+n <- 1000
+x <- runif(n, min=-1, max=1)
+y <- runif(n, min=-1, max=1)
+z <- f(x, y) + rnorm(n, sd = 0.01)
+dat <- data.frame(x=x, y=y, z=z)
+
+obj8 <- LMMsolve(fixed = z ~ 1,
+                spline = ~spl2D(x1 = x, x2 = y, x1lim = c(-1,1), x2lim=c(-1,1),
+                                nseg = c(25, 25)),
+                data = dat)
+
+newdat8 <- expand.grid(x = seq(-1, 1, length = 10),
+                      y = seq(-1, 1, length = 10))
+
+pred8 <- predict(obj8, newdata = newdat8, deriv = "x")
+expect_equivalent_to_reference(pred8, "pred8")
+
+pred9 <- predict(obj8, newdata = newdat8, deriv = "y")
+expect_equivalent_to_reference(pred9, "pred9")
+
+obj9 <- LMMsolve(fixed = z ~ 1,
+                 spline = ~spl1D(x = x, nseg = 20,xlim=c(-1,1)) +
+                           spl1D(x = y, nseg = 20,xlim=c(-1,1)),
+                 data = dat)
+expect_error(predict(obj9, newdat = newdat8, deriv = "x"),
+             "Derivatives for multiple splines not implemented yet")
+
+